% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_master.R
\name{app_master}
\alias{app_master}
\title{R6 master class (Controller) for your shiny application.}
\description{
app_master is the master class used across all modules of the the shiny app. It Stores all datasets needed in the applications
as tibbles. These tibbles are stored in \code{reactiveValues} \code{rvals}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{params}}{initialization parameters}

\item{\code{reactive_vals}}{reactiveValues instance that stores all reactive values needed in app}

\item{\code{static}}{values, non-reactive static datasets needed in the application. Note you can only pre-load these}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{static}}{values, non-reactive static datasets needed in the application. Note you can only pre-load these}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{app_master$new()}}
\item \href{#method-preload_master}{\code{app_master$preload_master()}}
\item \href{#method-add_master_data_row}{\code{app_master$add_master_data_row()}}
\item \href{#method-remove_dataset}{\code{app_master$remove_dataset()}}
\item \href{#method-replace_dataset_by_name}{\code{app_master$replace_dataset_by_name()}}
\item \href{#method-preload_master_with_config}{\code{app_master$preload_master_with_config()}}
\item \href{#method-dataset_names}{\code{app_master$dataset_names()}}
\item \href{#method-dataset_by_name}{\code{app_master$dataset_by_name()}}
\item \href{#method-data_by_index}{\code{app_master$data_by_index()}}
\item \href{#method-colnames_for_dataset}{\code{app_master$colnames_for_dataset()}}
\item \href{#method-prettynames_for_dataset}{\code{app_master$prettynames_for_dataset()}}
\item \href{#method-clone}{\code{app_master$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Standard R6 Initialize function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$new(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{the config yml driven params for initialization}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{app_master} object
Preload app_master with master_data `

Note the mdata value in reactiveValues (rvals) will be overwritten by what is provided
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preload_master"></a>}}
\if{latex}{\out{\hypertarget{method-preload_master}{}}}
\subsection{Method \code{preload_master()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$preload_master(master_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{master_data}}{mdata in rvals will be raplced with the mmaster_data provided
Preload app_master with csv files provided in config `

Note this creates new mdata overiding rvals}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_master_data_row"></a>}}
\if{latex}{\out{\hypertarget{method-add_master_data_row}{}}}
\subsection{Method \code{add_master_data_row()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$add_master_data_row(row)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-remove_dataset}{}}}
\subsection{Method \code{remove_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$remove_dataset(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace_dataset_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-replace_dataset_by_name}{}}}
\subsection{Method \code{replace_dataset_by_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$replace_dataset_by_name(dataset_name, replace_with)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preload_master_with_config"></a>}}
\if{latex}{\out{\hypertarget{method-preload_master_with_config}{}}}
\subsection{Method \code{preload_master_with_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$preload_master_with_config()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self object
access dataset names as list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset_names"></a>}}
\if{latex}{\out{\hypertarget{method-dataset_names}{}}}
\subsection{Method \code{dataset_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$dataset_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the names of datasets
search for a tibble based on dataset_name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-dataset_by_name}{}}}
\subsection{Method \code{dataset_by_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$dataset_by_name(dataset_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_name}}{the name of the dataset to lookup}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the mapped dataset in tibble format
search for a tibble based on index in mdata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_by_index"></a>}}
\if{latex}{\out{\hypertarget{method-data_by_index}{}}}
\subsection{Method \code{data_by_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$data_by_index(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{the row index of the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the mapped dataset in tibble format
get colnames for a dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-colnames_for_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-colnames_for_dataset}{}}}
\subsection{Method \code{colnames_for_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$colnames_for_dataset(dataset_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_name}}{the name of the dataset to lookup}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
characted list of colnames
get pretty colnames  for a dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettynames_for_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-prettynames_for_dataset}{}}}
\subsection{Method \code{prettynames_for_dataset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$prettynames_for_dataset(dataset_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_name}}{the name of the dataset to lookup}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
characted list of colnames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
