% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_mod.R
\name{corr_mod_onLoad}
\alias{corr_mod_onLoad}
\title{onLoad for corr_mod}
\usage{
corr_mod_onLoad(control, params)
}
\arguments{
\item{control}{the controller  \code{app_master}}

\item{params}{for the mod}
}
\description{
A shiny spring Module.
}
