% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_mod.R
\name{corr_mod_ui}
\alias{corr_mod_ui}
\title{ui_function for corr_mod}
\usage{
corr_mod_ui(id, control, params)
}
\arguments{
\item{id}{the id}

\item{control}{the controller  \code{app_master}}

\item{params}{for the mod}
}
\description{
A shiny spring Module UI function
}
