% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload_mod.R
\name{dataset_tbl}
\alias{dataset_tbl}
\title{Create the reactable table that shows the datasets}
\usage{
dataset_tbl(input, output, session, tbl_df)
}
\arguments{
\item{input}{standard shiny module parameter}

\item{output}{standard shiny module parameter}

\item{session}{standard shiny module parameter}

\item{tbl_df}{the data.frame that will be drawn with reatable}
}
\description{
The table shows the master mdata
}
