% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_connection_mod.R
\name{datasets_connection_ui}
\alias{datasets_connection_ui}
\title{datasets connection UI Function}
\usage{
datasets_connection_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
Connect various datasets to sweetmods for use
}
