% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_mod.R
\name{dummy_mod_onLoad}
\alias{dummy_mod_onLoad}
\title{onLoad for Dummy Mod}
\usage{
dummy_mod_onLoad(control, params)
}
\arguments{
\item{control}{the controller  \code{app_master}}

\item{params}{for the mod}
}
\description{
A shiny Module.
}
