% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_mod.R
\name{dummy_mod_ui}
\alias{dummy_mod_ui}
\title{UI Function for dummy mod}
\usage{
dummy_mod_ui(id, control, params)
}
\arguments{
\item{id}{Internal parameters for \code{shiny}}
}
\description{
A shiny Module.
}
