% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisse_wrapper.R
\name{esquisse_wrapper_onLoad}
\alias{esquisse_wrapper_onLoad}
\title{on Load}
\usage{
esquisse_wrapper_onLoad(control, params)
}
\arguments{
\item{control}{the app_master controller}

\item{params}{the params from config file}
}
\description{
A shiny Module.
}
