% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisse_wrapper.R
\name{esquisse_wrapper_server}
\alias{esquisse_wrapper_server}
\title{esquisse wrapper Server Functions}
\usage{
esquisse_wrapper_server(id, master)
}
\arguments{
\item{id}{standard shiny param}

\item{master}{the reference the app_master R6Class}
}
\description{
Calls esquisse module based on the selection made in radios input
}
