% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisse_wrapper.R
\name{esquisse_wrapper_ui}
\alias{esquisse_wrapper_ui}
\title{esquisse_wrapper UI Function}
\usage{
esquisse_wrapper_ui(id, control, params)
}
\arguments{
\item{id}{Internal parameters for \code{shiny}}

\item{params}{the params from config file}

\item{the}{app_master controller}
}
\description{
A shiny Module.
}
