% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload_mod.R
\name{file_upload_server}
\alias{file_upload_server}
\title{data_upload Server Functions}
\usage{
file_upload_server(id, master)
}
\arguments{
\item{id}{standard shiny server signature}

\item{master}{the \code{app_master} reference for this module}
}
\description{
data_upload Server Functions
}
