% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload_mod.R
\name{file_upload_ui}
\alias{file_upload_ui}
\title{data_upload UI Function}
\usage{
file_upload_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A File Data Upload Module.
}
