% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetutil.R
\name{load_built_ts_as_tibble}
\alias{load_built_ts_as_tibble}
\title{load a builtin datasource as tibble}
\usage{
load_built_ts_as_tibble(ds_name)
}
\arguments{
\item{ds_name}{}
}
\value{
data frame for the built in dataset
}
\description{
load a builtin datasource as tibble
}
