% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_manager.R
\name{load_row}
\alias{load_row}
\title{Loads a Row based on file type}
\usage{
load_row(name, ds_params, controller)
}
\arguments{
\item{name}{the name of dataset}

\item{ds_params}{the params specific to the ds}

\item{controller}{app_master}
}
\value{
row a Row object
}
\description{
Loads a Row based on file type
}
