% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_registry.R
\name{mod_registry}
\alias{mod_registry}
\title{R6  class for Modules Management for the shinyspring app}
\description{
mod_registry is the call you cn
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mod_names}}{initialization parameters}

\item{\code{master_params}}{initialization parameters}

\item{\code{mod_params}}{initialization parameters}

\item{\code{registry}}{get mod names from config files}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mod_registry$new()}}
\item \href{#method-mods_names}{\code{mod_registry$mods_names()}}
\item \href{#method-params_for_mod}{\code{mod_registry$params_for_mod()}}
\item \href{#method-validate_params}{\code{mod_registry$validate_params()}}
\item \href{#method-print}{\code{mod_registry$print()}}
\item \href{#method-clone}{\code{mod_registry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_registry$new(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mods_names"></a>}}
\if{latex}{\out{\hypertarget{method-mods_names}{}}}
\subsection{Method \code{mods_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_registry$mods_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
characted list of mod_names
get sub params for a given mod_name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-params_for_mod"></a>}}
\if{latex}{\out{\hypertarget{method-params_for_mod}{}}}
\subsection{Method \code{params_for_mod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_registry$params_for_mod(mod_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod_name}}{the mod_name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list ofparams
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_params"></a>}}
\if{latex}{\out{\hypertarget{method-validate_params}{}}}
\subsection{Method \code{validate_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_registry$validate_params()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_registry$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_registry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
