% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetutil.R
\name{new_dqmaster_row}
\alias{new_dqmaster_row}
\title{creates new dq_master row}
\usage{
new_dqmaster_row(sr_num, ds_name, ds)
}
\arguments{
\item{ds_name}{the unique name for the dataset you wand to refer it with to access it in the application}

\item{ds}{the dataset}

\item{srnum}{the serial number for the row. Index}
}
\value{
row the row
}
\description{
creates new dq_master row
}
