% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetutil.R
\name{parse_preloads_in_config}
\alias{parse_preloads_in_config}
\title{Parse preload files named in params (config).}
\usage{
parse_preloads_in_config(value, sep = ";")
}
\arguments{
\item{value}{the string that needs to be parsed}

\item{sep}{the seperator , default is ;}
}
\description{
In config file the attribute file_preloads will have a value like
file1 ; /home/file2 . This function will parse this into the file list
}
