% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_mod.R
\name{rmd_mod_onLoad}
\alias{rmd_mod_onLoad}
\title{onLoad for rmd_mod}
\usage{
rmd_mod_onLoad(control, params)
}
\arguments{
\item{control}{the controller  \code{app_master}}

\item{params}{for the mod}
}
\description{
A shiny spring Module.
}
