% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_mod.R
\name{rmd_mod_server}
\alias{rmd_mod_server}
\title{server_function for rmd_mod}
\usage{
rmd_mod_server(id, control, params)
}
\arguments{
\item{id}{the id}

\item{control}{the controller  \code{app_master}}

\item{params}{for the mod}
}
\description{
A shiny spring Module UI function
}
