% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_eda.R
\name{smart_eda_server}
\alias{smart_eda_server}
\title{EDA Server components}
\usage{
smart_eda_server(id, master)
}
\arguments{
\item{id}{standard shiny server signature}

\item{master}{the \code{app_master} reference for this module}
}
\description{
EDA Server components
}
