% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_datamod.R
\name{sweet_datamod_ui}
\alias{sweet_datamod_ui}
\title{data_upload UI Function}
\usage{
sweet_datamod_ui(id, control)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\description{
A File Data Upload Module.
}
