% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_mod.R
\name{use_rmd_mod}
\alias{use_rmd_mod}
\title{Created rmd mod dependencies and displays the properties needed for rmd mod}
\usage{
use_rmd_mod(
  registry,
  rmd_file = "intro.Rmd",
  template = "templates/rmd_mod_t1.mst"
)
}
\arguments{
\item{registry}{the registry}

\item{rmd_file}{optional}

\item{template}{optional}
}
\description{
Created rmd mod dependencies and displays the properties needed for rmd mod
}
