% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetutil.R
\name{ymlon_to_params}
\alias{ymlon_to_params}
\title{used to get sub params for a given mod_name}
\usage{
ymlon_to_params(obj_name, master_params)
}
\arguments{
\item{mod_name}{the mod_name}

\item{the}{master param}
}
\value{
list of params
}
\description{
used to get sub params for a given mod_name
}
