% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_selection.R
\name{column_selection_mod_server}
\alias{column_selection_mod_server}
\title{server_function for column_selection_mod}
\usage{
column_selection_mod_server(
  id,
  control,
  dq_v,
  dataset_names,
  filter_types = c("everything")
)
}
\arguments{
\item{id}{the id}

\item{control}{the controller  \code{app_master}}

\item{params}{for the mod}
}
\description{
A shiny spring Module UI function
}
