% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_selection.R
\name{column_selection_mod_ui}
\alias{column_selection_mod_ui}
\title{ui_function for column_selection_mod}
\usage{
column_selection_mod_ui(id, dataset_names)
}
\arguments{
\item{id}{the id}

\item{control}{the controller  \code{app_master}}

\item{params}{for the mod}
}
\description{
A shiny spring Module UI function
}
