% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisse_wrapper.R
\name{esquisse_wrapper_server}
\alias{esquisse_wrapper_server}
\title{esquisse_wrapper Server Function}
\usage{
esquisse_wrapper_server(id, control, params)
}
\arguments{
\item{id}{Internal parameters for \code{shiny}}

\item{control}{the app_master controller}

\item{params}{the params from config file}
}
\description{
A shiny Module.
}
