% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esquisse_wrapper.R
\name{esquisse_wrapper_server}
\alias{esquisse_wrapper_server}
\title{esquisse_wrapper Server Function}
\usage{
esquisse_wrapper_server(id, control, params = NULL)
}
\arguments{
\item{id}{Internal parameters for \code{shiny}}
}
\description{
A shiny Module.
}
