% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetutil.R
\name{load_tar_as_tibble}
\alias{load_tar_as_tibble}
\title{load a target and put it into a tibble.
Note : we clear up the tar_load value before reutrnignt he dataframe to free up the memory}
\usage{
load_tar_as_tibble(tar_name)
}
\description{
load a target and put it into a tibble.
Note : we clear up the tar_load value before reutrnignt he dataframe to free up the memory
}
